requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.0.0
